More shuffling of logic into protocols and out of plugin
